+++ src/ops/launcher/gqa_attention_prefill.cu
@@ -54,7 +54,7 @@ void gqa_attention_prompt_attention_launch_for(const Tensor& q, const Tensor& po
         <<<attention_grid, kGqaPrefillSimtThreads, 0, stream>>>(
             static_cast<const __nv_bfloat16*>(q.data),
             static_cast<const __nv_bfloat16*>(cache_k.data),
-            static_cast<const __nv_bfloat16*>(cache_v.data), metadata,
+            static_cast<const __half*>(cache_v.data), metadata,
             static_cast<const std::int32_t*>(positions.data), scale,
             static_cast<__nv_bfloat16*>(out.data), tokens);
     CUDA_CHECK(cudaGetLastError());
@@ -91,7 +91,7 @@ void gqa_attention_prompt_attention_launch_for(const Tensor& q, const Tensor& po
             <<<attention_grid, kGqaPrefillThreads, kGqaPrefillSmemBytes, stream>>>(
                 static_cast<const __nv_bfloat16*>(q.data),
                 static_cast<const __nv_bfloat16*>(cache_k.data),
-                static_cast<const __nv_bfloat16*>(cache_v.data), metadata,
+                static_cast<const __half*>(cache_v.data), metadata,
                 static_cast<const std::int32_t*>(positions.data), scale,
                 static_cast<__nv_bfloat16*>(out.data), tokens);
     }
@@ -154,7 +154,7 @@ void gqa_kv_append_launch_for(const Tensor& k, const Tensor& v, const Tensor& po
                                                static_cast<const __nv_bfloat16*>(v.data),
                                                static_cast<const std::int32_t*>(positions.data),
                                                metadata, static_cast<__nv_bfloat16*>(cache_k.data),
-                                               static_cast<__nv_bfloat16*>(cache_v.data), tokens);
+                                               static_cast<__half*>(cache_v.data), tokens);
         CUDA_CHECK(cudaGetLastError());
     }
 }
